implement mini-agent against a struct. allow for the injection of a custom trace processor
