+++ serverless/Cargo.toml
@@ -1,12 +1,9 @@
 [package]
-name = "serverless_mini_agent"
+name = "trace_mini_agent"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-hyper = { version = "0.14", features = ["full"] }
-tokio = { version = "1", features = ["full"] }
-serde_json = "1.0.87"
-prost = "0.11.6"
+datadog-trace-mini-agent = { path = "../trace-mini-agent" }
 datadog-trace-protobuf = { path = "../trace-protobuf" }
 datadog-trace-utils = { path = "../trace-utils" }
