+++ trace-mini-agent/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "datadog-trace-mini-agent"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+hyper = { version = "0.14", features = ["full"] }
+tokio = { version = "1", features = ["full"] }
+serde_json = "1.0.87"
+async-trait = "0.1.68"
+dyn-clone = "1.0.11"
+datadog-trace-protobuf = { path = "../trace-protobuf" }
+datadog-trace-utils = { path = "../trace-utils" }
