+++ trace-utils/Cargo.toml
@@ -8,5 +8,6 @@ edition = "2021"
 curl = "0.4.44"
 hyper = { version = "0.14", features = ["full"] }
 serde = { version = "1.0.145", features = ["derive"] }
+prost = "0.11.6"
 rmp-serde = "1.1.1"
 datadog-trace-protobuf = { path = "../trace-protobuf" }
