+++ trace-utils/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod trace_sender;
-pub mod trace_type_converter;
+pub mod trace_utils;
